+++ QueenBeeParser/Zlib/InfBlocks.cs
@@ -143,7 +143,7 @@ namespace Rebex.IO.Compression
 				p = z.next_in_index; n = z.avail_in; b = bitb; k = bitk;
 			}
 			{
-				q = write; m = (int) (q < read?read - q - 1:end - q);
+				q = write; m = q < read ? read - q - 1 : end - q;
 			}
 			
 			// process input based on current state
@@ -173,7 +173,7 @@ namespace Rebex.IO.Compression
 							b |= (z.next_in[p++] & 0xff) << k;
 							k += 8;
 						}
-						t = (int) (b & 7);
+						t = b & 7;
 						last = t & 1;
 						
 						switch (SupportClass.URShift(t, 1))
@@ -285,16 +285,16 @@ namespace Rebex.IO.Compression
 						{
 							if (q == end && read != 0)
 							{
-								q = 0; m = (int) (q < read?read - q - 1:end - q);
+								q = 0; m = q < read ? read - q - 1 : end - q;
 							}
 							if (m == 0)
 							{
 								write = q;
 								r = inflate_flush(z, r);
-								q = write; m = (int) (q < read?read - q - 1:end - q);
+								q = write; m = q < read ? read - q - 1 : end - q;
 								if (q == end && read != 0)
 								{
-									q = 0; m = (int) (q < read?read - q - 1:end - q);
+									q = 0; m = q < read ? read - q - 1 : end - q;
 								}
 								if (m == 0)
 								{
@@ -579,7 +579,7 @@ namespace Rebex.IO.Compression
 						codes.free(z);
 						
 						p = z.next_in_index; n = z.avail_in; b = bitb; k = bitk;
-						q = write; m = (int) (q < read?read - q - 1:end - q);
+						q = write; m = q < read ? read - q - 1 : end - q;
 						
 						if (last == 0)
 						{
@@ -592,7 +592,7 @@ namespace Rebex.IO.Compression
 					case DRY: 
 						write = q;
 						r = inflate_flush(z, r);
-						q = write; m = (int) (q < read?read - q - 1:end - q);
+						q = write; m = q < read ? read - q - 1 : end - q;
 						if (read != write)
 						{
 							bitb = b; bitk = k;
@@ -665,7 +665,7 @@ namespace Rebex.IO.Compression
 			q = read;
 			
 			// compute number of bytes to copy as far as end of window
-			n = (int) ((q <= write?write:end) - q);
+			n = (q <= write ? write : end) - q;
 			if (n > z.avail_out)
 				n = z.avail_out;
 			if (n != 0 && r == Z_BUF_ERROR)
