+++ QueenBeeParser/Zlib/Inflate.cs
@@ -372,7 +372,7 @@ namespace Rebex.IO.Compression
 			return Z_OK;
 		}
 		
-		private static byte[] mark = new byte[]{(byte) 0, (byte) 0, (byte) 0xff, (byte) 0xff};
+		private static byte[] mark = new byte[]{ 0, 0, 0xff, 0xff };
 		
 		internal int inflateSync(ZStream z)
 		{
